+++ lib/server/trpc/routers/postRouter.ts
@@ -3,8 +3,8 @@ import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 import { decodeCursor, encodeCursor } from "~/lib/server/cursor";
 import { prisma } from "~/lib/server/prisma";
+import { revalidator } from "~/lib/server/revalidator";
 import { loggedProcedure, trpc } from "~/lib/server/trpc/trpc";
-import { revalidator } from "../../revalidator";
 
 const DEFAULT_PER_PAGE = 20;
